fix typedoc generation, move types to new file
