+++ src/library/client/BridgeClient.ts
@@ -1,17 +1,17 @@
-import { Display, tryParseDisplay } from "./components/displays"
-import { sendMessage } from "./components/endpoints"
-import { tryEnterOrchestration, tryExitOrchestration } from "./components/orchestration"
-import { BridgeEventSource } from "./components/eventsource"
-import { Playlist } from "./playlists/playlist"
-import { HologramType } from "./components/hologram"
-import { BridgeEventMap } from "./schemas/schema.events"
-import { HologramParamMap } from "./schemas/schema.parameters"
-import * as schema from "./schemas/schema.responses"
+import { Display, tryParseDisplay } from "../components/displays"
+import { sendMessage } from "../components/endpoints"
+import { tryEnterOrchestration, tryExitOrchestration } from "../components/orchestration"
+import { BridgeEventSource } from "../components/eventsource"
+import { Playlist } from "../playlists/playlist"
+import { HologramType } from "../components/hologram"
+import { BridgeEventMap } from "../schemas/schema.events"
+import { HologramParamMap } from "../schemas/schema.parameters"
+import * as schema from "../schemas/schema.responses"
 import { z } from "zod"
-import { Fallback } from "./components/fallback"
-import { NewItemPlayingMessageHandler } from "./components/messageHandler"
-import { BridgeVersion } from "./components"
-import { parseBridgeVersion } from "./utilities/general.utils"
+import { Fallback } from "../components/fallback"
+import { NewItemPlayingMessageHandler } from "../components/messageHandler"
+import { BridgeVersion } from "../components/types"
+import { parseBridgeVersion } from "../utilities/general.utils"
 
 export class BridgeClient {
 	/** The name of the current orchestration */
