+++ src/library/index.ts
@@ -1,8 +1,9 @@
-export * from "./client"
+export * from "./client/BridgeClient"
 export * from "./playlists/playlist"
 export * from "./playlists/playlistItems"
 export * from "./components/messageHandler"
 export * from "./components/hologram"
+export * from "./components/types"
 export * from "./schemas/schema.hologram"
 export * from "./schemas/schema.parameters"
 export * as defaults from "./schemas/defaults"
