+++ src/react-app/tsconfig.json
@@ -5,5 +5,5 @@
 		"jsx": "react-jsx",
 		"outDir": "../../app"
 	},
-	"include": ["**/*.ts", "**/*.tsx", "../library/**/*.ts", "../library/**/*/d.ts"]
+	"include": ["**/*.ts", "**/*.tsx", "../library/**/*.ts"]
 }
